requirements.txt
click==8.1.7
pydantic==2.13.0
beautifulsoup4==4.12.2
jinja2==3.1.2
python-dotenv==1.0.0
openai>=1.3.0
requests==2.31.0
google-play-scraper==1.2.7
